requirements/base.txt
paho-mqtt==1.6.1
HAP-python[QRCode]==4.4.0